pyproject.toml
[tool.poetry]
name = "tgrpc"
version = "0.1.1"
description = ""
authors = ["Vladislav Trofimenko <slashcooperlive@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"
websockets = "^11.0.3"
simplejson = "^3.19.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
